+++ sydtest-persistent-postgresql/default.nix
@@ -6,14 +6,16 @@
 }:
 mkDerivation {
   pname = "sydtest-persistent-postgresql";
-  version = "0.5.1.0";
+  version = "0.6.0.0";
   src = ./.;
   libraryHaskellDepends = [
     base directory filepath monad-logger mtl persistent-postgresql
     postgres-options postgresql-simple random sydtest
     sydtest-persistent temporary text time tmp-postgres typed-process
   ];
-  testHaskellDepends = [ base persistent postgresql-simple sydtest ];
+  testHaskellDepends = [
+    base persistent postgresql-simple sydtest tmp-postgres
+  ];
   testToolDepends = [ sydtest-discover ];
   homepage = "https://github.com/NorfairKing/sydtest#readme";
   description = "An persistent-postgresql companion library for sydtest";